Documentation, basic LED driver, reorganize code
Some documentation has been added in the form a PDF that describes
eForth (for the x86 under MS-DOS, but the concepts are the same/similar
to those used in 'embed') and a project that implements a Morse code
decoder for an ATTiny.

The basics for the LED sensing hack have been added, it appears to be
partially working but there is more to do in order to get this working
as a bidirectional communications mechanism.

Some of the code has been reorganized/improved generally.